So, progress on unfinished stuff.

+++ 3DRadSpace/3DRadSpaceDll/TextPrint.cs
@@ -122,7 +122,7 @@ namespace _3DRadSpaceDll
         /// <summary>
         /// Depth layer.
         /// </summary>
-        public int Layer { get; set; }
+        public float Layer { get; set; }
 
         /// <summary>
         /// Loads the font.
